trim to workflow, pull out rubric process
